+++ src/NotSpecification.php
@@ -8,7 +8,7 @@ namespace Maartenpaauw\Specifications;
  * @template TCandidate
  * @extends CompositeSpecification<TCandidate>
  */
-class NotSpecification extends CompositeSpecification
+final class NotSpecification extends CompositeSpecification
 {
     /**
      * @param Specification<TCandidate> $specification