+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to `laravel-specification-pattern` will be documented in this file.
 
-## Unreleased
+## [v2.0.0] - 2023-03-21
+
+### Added
+
+- Support from Laravel 10.
 
 ### Changed
 
@@ -12,6 +16,11 @@ All notable changes to `laravel-specification-pattern` will be documented in thi
   - `NotSpecification`.
   - `OrSpecification`.
   - `VerboseSpecification`.
+
+### Removed
+
+- Support for PHP 8.
+- Support for Laravel 8.
 
 ## [v1.1.0] - 2022-03-07
 
@@ -23,5 +32,6 @@ All notable changes to `laravel-specification-pattern` will be documented in thi
 
 - initial release
 
+[v2.0.0]: https://github.com/maartenpaauw/laravel-specification-pattern/compare/v1.1.0...v2.0.0
 [v1.1.0]: https://github.com/maartenpaauw/laravel-specification-pattern/compare/v1.0.0...v1.1.0
 [v1.0.0]: https://github.com/maartenpaauw/laravel-specification-pattern/releases/tag/v1.0.0